fix bug #1290 that displayed simulation tab for all layers (#1304)
* fix bug #1290 that displayed simulation tab for all layers

* fix bug #1290 that displayed simulation tab for all layers

---------
